Bunch of refactoring and other code cleanup and warning removal
Added a sendPacket function to freeemscomms.h/cpp, which should
facilitate a OAOO approach. It greatly reduced the number of repeated
code within the thread loop. It also allowed for significant
expandability in the future in terms of types of arguments to send
along with messages and how they are parsed into byte arrays.

+++ src/datafield.cpp
@@ -21,7 +21,7 @@
 DataField::DataField()
 {
 }
-DataField::DataField(QString shortname,QString description,int offset,int size,float div,float min,float max,float addoffset,bool isFlags,int bit)
+DataField::DataField(QString shortname,QString description,int offset,int size,double div,double min,double max,double addoffset,bool isFlags,int bit)
 {
 	m_offset = offset;
 	m_size = size;
